requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
bleach==3.1.0
canoser==0.5.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
docutils==0.15.2
grpcio==1.23.0
grpcio-tools==1.23.0
idna==2.8
importlib-metadata==0.21
mnemonic==0.18
more-itertools==7.2.0
packaging==19.1
pbkdf2==1.3
pkginfo==1.5.0.1
pluggy==0.13.0
protobuf==3.9.1
py==1.8.0
pycparser==2.19
Pygments==2.4.2
PyNaCl==1.3.0
pyparsing==2.4.2
pysha3==1.0.2
pytest==5.1.2
readme-renderer==24.0
requests==2.22.0
requests-toolbelt==0.9.1
six==1.12.0
toml==0.10.0
tqdm==4.35.0
twine==1.14.0
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
zipp==0.6.0
